switched from jimp to sharp

+++ packages/images/src/index.ts
@@ -1,42 +1,58 @@
-/* eslint-disable no-continue */
-import Jimp from 'jimp';
-import {RaggedyObject, TWorker} from '@lollygag/core';
+/* eslint-disable no-param-reassign */
+import {fullExtname, TWorker} from '@lollygag/core';
+import sharp, {GifOptions, PngOptions, JpegOptions} from 'sharp';
 import {existsSync, mkdirSync, readFileSync, Stats, writeFileSync} from 'fs';
 import {writeFile} from 'fs/promises';
+import {basename, join} from 'path';
 
 export interface IImagesOptions {
-    imageCompressorOptions?: RaggedyObject;
+    gifOptions?: GifOptions;
+    pngOptions?: PngOptions;
+    jpegOptions?: JpegOptions;
+    widths?: number[];
+}
+
+function generateFilename(
+    path: string,
+    id: number | string,
+    quality: number | false
+) {
+    const fileExt = fullExtname(path);
+    const fileName = basename(path, fileExt);
+
+    if(quality) return `${fileName}-${id}-q${quality}${fileExt}`;
+
+    return `${fileName}-${id}${fileExt}`;
 }
 
 export default function images(options?: IImagesOptions): TWorker {
     return async function imagesWorker(this: TWorker, files): Promise<void> {
         if(!files) return;
 
-        const metaFile = '.meta/lollygag-images.json';
+        const {gifOptions, pngOptions, jpegOptions, widths} = options ?? {};
 
-        if(!existsSync('.meta/')) mkdirSync('.meta');
+        const metaFile = '.images/meta.json';
 
-        if(!existsSync(metaFile)) {
-            writeFileSync(metaFile, '{}');
-        }
+        if(!existsSync('.images/')) mkdirSync('.images');
+        if(!existsSync(metaFile)) writeFileSync(metaFile, '{}');
 
         interface IImagesMeta {
             [path: string]: {
-                birthtime: Stats['birthtime'];
+                birthtimeMs: Stats['birthtimeMs'];
             };
         }
-
-        const meta: IImagesMeta = JSON.parse(
-            readFileSync(metaFile, {encoding: 'utf-8'}) ?? '{}'
-        );
 
         const promises = files.map(async(file) => {
             if(!file.mimetype.startsWith('image')) return;
 
+            const meta: IImagesMeta = JSON.parse(
+                readFileSync(metaFile, {encoding: 'utf-8'}) ?? '{}'
+            );
+
             if(meta[file.path] && file.stats) {
                 if(
-                    new Date(meta[file.path].birthtime)
-                    >= new Date(file.stats.birthtime)
+                    new Date(meta[file.path].birthtimeMs)
+                    >= new Date(file.stats.birthtimeMs)
                 ) {
                     return;
                 }
@@ -44,21 +60,46 @@ export default function images(options?: IImagesOptions): TWorker {
 
             console.log(`Processing ${file.path}...`);
 
-            await Jimp.read(file.path).then((img) => {
-                img.quality(75).resize(1200, Jimp.AUTO).write(file.path);
+            const img = sharp(file.path);
 
-                if(file.stats) {
-                    meta[file.path] = {
-                        birthtime: file.stats.birthtime,
-                    };
-                }
+            let quality: number | false = false;
+
+            switch(file.mimetype) {
+                case 'image/gif':
+                    img.gif(gifOptions);
+                    break;
+                case 'image/png':
+                    quality = pngOptions?.quality ?? 100;
+                    img.png({quality, ...pngOptions});
+                    break;
+                case 'image/jpeg':
+                    quality = jpegOptions?.quality ?? 80;
+                    img.jpeg({quality, ...jpegOptions});
+                    break;
+                default:
+            }
+
+            await img.toFile(
+                join('.images', generateFilename(file.path, 'full', quality))
+            );
+
+            widths?.forEach((width) => {
+                img.resize(width).toFile(
+                    join('.images', generateFilename(file.path, width, quality))
+                );
             });
 
+            if(file.stats) {
+                meta[file.path] = {
+                    birthtimeMs: file.stats.birthtimeMs,
+                };
+            }
+
             console.log(`Processing ${file.path}... done!`);
+
+            await writeFile(metaFile, JSON.stringify(meta, null, 2));
         });
 
         await Promise.all(promises);
-
-        await writeFile(metaFile, JSON.stringify(meta));
     };
 }
